fix(ci): make gen_vectors a runnable example to unblock clippy
`crates/zk-soroban/src/bin.rs` (the `gen_vectors` G2-coordinate printer added
in #271) was declared as a `[[bin]]` but its dependencies — `num-bigint`,
and the full-featured `ark-bn254`/`ark-ec` — live in `[dev-dependencies]`.
Binaries cannot see dev-dependencies, so `cargo clippy --all-targets`
(the "Strict Linting" CI step) failed for every PR and push to main:

    error[E0432]: unresolved import `num_bigint`
    error: unused variable: `y_c0` / `y_c1`   (their println! was commented out)

The dev-dependency placement shows this was always meant to be an example,
not a shipping binary. Fix:

- Move `src/bin.rs` -> `examples/gen_vectors.rs` and drop the `[[bin]]` entry.
  Examples *can* use dev-dependencies, so `num-bigint` and the full ark crates
  now resolve, and the example is excluded from the no_std/WASM contract build.
- Print all four G2 coordinates (the Y c0/c1 lines were previously commented
  out, leaving `y_c0`/`y_c1` unused). Output matches the hardcoded vectors in
  `pairing.rs`.
- Add a short doc header and the run command.

Verified locally: `cargo clippy --all-targets --all-features -- -D warnings`,
`cargo fmt --check`, verifier-sample WASM build (within 64 KB), and
`cargo run -p zk-soroban --example gen_vectors` all pass.
