+++ crates/zk-soroban/Cargo.toml
@@ -7,10 +7,6 @@ license.workspace = true
 [lib]
 crate-type = ["lib"]
 
-[[bin]]
-name = "gen_vectors"
-path = "src/bin.rs"
-
 [dependencies]
 # Inherited from workspace (no_std enabled)
 soroban-sdk = { workspace = true, features = ["hazmat-crypto"]}
